fix(ci): drop stray oobe_fail.txt and ignore nightly dir in default collection
Two independent CI failures on PR #266:

1. check-version flagged a mandatory version bump because oobe_fail.txt
   (a Playwright failure dump accidentally committed in d0567e6) sits
   at the repo root and is not covered by the workflow skip regex.
   Deleted the file.

2. The unit-test job runs `pytest tests/ --ignore=tests/e2e`; the
   nightly modules import playwright at module scope, which crashes
   collection before tests/nightly/conftest.py's modifyitems skip can
   fire. Registered --run-nightly in tests/conftest.py and added a
   pytest_ignore_collect hook there so the nightly tree is skipped
   entirely unless the flag or NIGHTLY env var is set. Removed the
   duplicate pytest_addoption from tests/nightly/conftest.py.

Verified locally: `pytest tests/ --co` collects 1329 unit tests with
no ImportError, `pytest tests/nightly --run-nightly --co` still
sees the full 55-test nightly suite.
